requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
django-ckeditor==5.3.0
django-disqus==0.5
django-embed-video==1.1.2
django-js-asset==0.1.1
django-taggit==0.22.1
gunicorn==19.7.1
idna==2.6
Markdown==2.6.9
mock==2.0.0
olefile==0.44
pbr==3.1.1
Pillow==4.2.1
psycopg2==2.7.3.1
pytz==2017.2
requests==2.18.4
six==1.10.0
urllib3==1.22
whitenoise==3.3.1
